Initial work on oss krakend config (#1)
This is my initial work on the krakend config including the endpoints for the load balancer api. This work is not complete, but I'm PRing as a base to work off of as others start to work on it. Any or all of this config is subject to change or to the trash bin.